Vastly simplify project structure and bug fixes
