+++ .gitignore
@@ -36,3 +36,5 @@ neutralino.js
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+
+.vscode
