+++ public/index.html
@@ -9,8 +9,7 @@
       name="description"
       content="Web site created using create-react-app"
     />
-    <script src="http://localhost:50388/__neutralino_globals.js"></script>
-    <script src="http://localhost:50388/neutralino.js"></script>
+    <script src="http://localhost:63998/__neutralino_globals.js"></script>
     <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
     <!--
       manifest.json provides metadata used when your web app is installed on a
