+++ src/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-
+import App from './components/App';
 import { init, events, app } from "@neutralinojs/lib"
-import App from './App';
+
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
@@ -13,5 +13,5 @@ root.render(
 
 init();
 events.on("windowClose", () => {
-  app.exit();
+  app.killProcess();
 });
